requirements.txt
pillow
requests
retry